requirements.txt
rembg
Pillow
easygui